Refactored streamlit app loading into separate component
